- Criado função que valida se o aluno tem 3 avaliações antes de adicionar uma nova avaliação - Criado tratativa que impede a deleção de uma turma que tenha numeroAlunos superior a 0 - Alterada a entidade BoletimDTO para visualização de informações no front

+++ src/main/java/br/com/hbsis/projetocursos/service/TurmaServiceImpl.java
@@ -72,8 +72,15 @@ public class TurmaServiceImpl implements TurmaService {
     }
 
     @Override
-    public void deleteById(int theId) {
-        turmaRepository.deleteById(theId);
+    public void deleteById(int theId) throws Exception {
+        Turma turma = findById(theId);
+        if(turma.getNumeroAlunos() > 0) {
+            throw new Exception("Não é possível deletar a turma, pois ela tem alunos, remova os alunos da turma antes de deletá-la");
+        }
+
+        else {
+            turmaRepository.deleteById(theId);
+        }
     }
 
     @Override
